requirements.txt
git+https://github.com/msuliot/package.utils.git
python-docx
PyPDF2
openpyxl
python-pptx
xlrd
pandas
PyMuPDF
nltk
numpy
requests
beautifulsoup4